# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: TrustFlow AI - Behavioral Authentication API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login with username and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Login successful or OTP required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResult"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/otp/verify:
    post:
      operationId: verifyOtp
      tags: [auth]
      summary: Verify OTP code
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OtpInput"
      responses:
        "200":
          description: OTP verified
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SessionInfo"
        "401":
          description: Invalid OTP
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/session:
    get:
      operationId: getSession
      tags: [auth]
      summary: Get current session info
      responses:
        "200":
          description: Session info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SessionInfo"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /behavior/event:
    post:
      operationId: recordBehaviorEvent
      tags: [behavior]
      summary: Record a behavioral event
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BehaviorEventInput"
      responses:
        "200":
          description: Event recorded and trust score updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TrustScoreUpdate"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
...
    LoginInput:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
        deviceFingerprint:
          type: string
        ipAddress:
          type: string
      required: [username, password]
    AuthResult:
      type: object
      properties:
        status:
          type: string
          enum: [success, otp_required, biometric_required, blocked]
        sessionToken:
          type: ["string", "null"]
        message:
          type: string
        userId:
          type: ["integer", "null"]
        username:
          type: ["string", "null"]
        riskLevel:
          type: string
          enum: [low, medium, high, critical]
      required: [status, message, riskLevel]
    OtpInput:
      type: object
      properties:
        sessionToken:
          type: string
        otp:
          type: string
      required: [sessionToken, otp]
    SessionInfo:
      type: object
      properties:
        userId:
          type: integer
        username:
          type: string
        sessionToken:
          type: string
        trustScore:
          type: integer
        riskLevel:
          type: string
          enum: [low, medium, high, critical]
        deviceFingerprint:
          type: ["string", "null"]
        ipAddress:
          type: ["string", "null"]
        loginTime:
          type: string
        isAuthenticated:
          type: boolean
      required: [userId, username, sessionToken, trustScore, riskLevel, loginTime, isAuthenticated]
    BehaviorEventInput:
      type: object
      properties:
        sessionToken:
          type: string
        eventType:
          type: string
          enum: [keystroke, mouse_move, click, scroll, page_focus, page_blur, idle]
        typingSpeed:
          type: ["number", "null"]
        keystrokeInterval:
          type: ["number", "null"]
        mouseVelocity:
          type: ["number", "null"]
        clickCount:
          type: ["integer", "null"]
        scrollSpeed:
          type: ["number", "null"]
        pageX:
          type: ["number", "null"]
        pageY:
          type: ["number", "null"]
        metadata:
          type: object
          additionalProperties: true
      required: [sessionToken, eventType]
    TrustScoreUpdate:
      type: object
      properties:
        trustScore:
          type: integer
        riskLevel:
          type: string
          enum: [low, medium, high, critical]
        riskFactors:
          type: array
          items:
            $ref: "#/components/schemas/RiskFactor"
        action:
          type: string
          enum: [allow, request_otp, request_biometric, block]
      required: [trustScore, riskLevel, riskFactors, action]
    RiskFactor:
      type: object
      properties:
        factor:
          type: string
        description:
          type: string
        severity:
          type: string
          enum: [low, medium, high]
        contribution:
          type: integer
      required: [factor, description, severity, contribution]
...