# lib/api-zod/src/generated/types/authResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * TrustFlow AI - Behavioral Authentication API
 * OpenAPI spec version: 0.1.0
 */
import type { AuthResultRiskLevel } from './authResultRiskLevel';
import type { AuthResultStatus } from './authResultStatus';

export interface AuthResult {
  status: AuthResultStatus;
  /** @nullable */
  sessionToken?: string | null;
  message: string;
  /** @nullable */
  userId?: number | null;
  /** @nullable */
  username?: string | null;
  riskLevel: AuthResultRiskLevel;
}