# lib/api-zod/src/generated/types/otpInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * TrustFlow AI - Behavioral Authentication API
 * OpenAPI spec version: 0.1.0
 */

export interface OtpInput {
  sessionToken: string;
  otp: string;
}