# lib/api-zod/src/generated/types/sessionInfo.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * TrustFlow AI - Behavioral Authentication API
 * OpenAPI spec version: 0.1.0
 */
import type { SessionInfoRiskLevel } from './sessionInfoRiskLevel';

export interface SessionInfo {
  userId: number;
  username: string;
  sessionToken: string;
  trustScore: number;
  riskLevel: SessionInfoRiskLevel;
  /** @nullable */
  deviceFingerprint?: string | null;
  /** @nullable */
  ipAddress?: string | null;
  loginTime: string;
  isAuthenticated: boolean;
}